refactor: aggregate-init providers, LangChain-aligned agents, LLM summarize, bugfixes
Breaking changes (no backward compat):
- Providers now use C++20 aggregate init: OpenAIChat{.model="...", .api_key=key}
- AgentExecutor renamed to agents::DeepAgent<LLM>, agent.hpp re-exports
- Log moved into AgentConfig.logger instead of separate constructor arg

New features:
- agents/ folder with DeepAgent, create_agent factory, bind() wrapper
- create_agent<deep_agent_tag>::create({.llm=..., .name=...}) single-param factory
- make_agent(llm, config) two-param shorthand
- LLMConfig::merge() for per-call kwargs via agent.invoke(input, LLMConfig{...})
- bind(llm, LLMConfig{...}) for pre-configured model wrappers
- LLM-based summarize middleware with chunking and default prompt
- Runnable | operator chaining (LangChain-style)

Bugfixes: HnswVectorStore move (use-after-free), StdioTransport move/partial-write,
embedding index bounds, DynamicTool typed() dangling ref, agent as_tool dangling this,
ToolCache JSON key ordering, log stream state leak, 1<<N UB, PII multi-part bypass,
Anthropic empty api_key guard, ModelConfig narrowing clamp, summarize dedup

Windows: _putenv_s, _popen/_pclose defines

New examples: 14_bind_and_kwargs, 15_llm_summarize, 16_deep_agent_custom

+++ include/tiny_agent/middleware/retry.hpp
@@ -25,7 +25,7 @@ struct Retry {
             catch (const APIError& e) {
                 if (attempt >= MaxRetries || !detail::is_retryable(e.status_code))
                     throw;
-                auto delay = BackoffMs * (1 << attempt);
+                auto delay = BackoffMs * (static_cast<long long>(1) << attempt);
                 log.warn("retry", "attempt " + std::to_string(attempt + 1)
                     + "/" + std::to_string(MaxRetries) + " failed (status="
                     + std::to_string(e.status_code) + "), retrying in "
@@ -45,7 +45,7 @@ inline MiddlewareFn retry(int max_retries = 3,
             catch (const APIError& e) {
                 if (attempt >= max_retries || !detail::is_retryable(e.status_code))
                     throw;
-                auto delay = backoff * (1 << attempt);
+                auto delay = backoff * (static_cast<long long>(1) << attempt);
                 log.warn("retry", "attempt " + std::to_string(attempt + 1)
                     + "/" + std::to_string(max_retries) + " failed (status="
                     + std::to_string(e.status_code) + "), retrying in "
